+++ protobuf/internalapi/generate.go
@@ -1,2 +1,4 @@
+//go:generate protoc --go_out=. --go-grpc_out=. ./base.proto
 //go:generate protoc --go_out=. --go-grpc_out=. ./dao.proto
+//go:generate protoc --go_out=. --go-grpc_out=. ./proposal.proto
 package internalapi